// src/app/interfaces/IUser.ts
export interface IUser {

      username: string; 
      first_name: string;
      last_name:  string;
      birth_date: string;
      institution: string;
      campus: string;
      student_code: string;
      student_carreer: string;  
      phone: string;
      email: string;
      password: string;
      confirm_password: string
}